Move initializeQB to its own file
