+++ frontend/src/metabase/query_builder/actions/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./core";
+export * from "./initializeQB";
